Separacion de modulos y agregar iconos
